+++ man-roxygen/ex-markov_mix_utils.R
@@ -15,3 +15,15 @@ cbind(
   as.data.frame(get_states_mat(markov_mix_ex)),
   as.data.frame(get_prob(markov_mix_ex))
 )
+
+# Subset 1 or more components
+markov_mix_ex[2L]
+markov_mix_ex[c(1L, 3L)]
+
+# Replace 1 or more components
+markov_mix_ex2 <- markov_mix_ex
+markov_mix_ex2[2L] <- runif(length(markov_mix_ex[["states"]])^(markov_mix_ex[["order"]] + 1L))
+print(markov_mix_ex2)
+markov_mix_ex3 <- markov_mix_ex
+markov_mix_ex3[c(1L, 3L)] <- matrix(runif(length(markov_mix_ex[["states"]])^(markov_mix_ex[["order"]] + 1L) * 2L), ncol = 2L)
+print(markov_mix_ex3)